fix: use the windows api function equivalent to rusage
