+++ releases/go/smithy-dafny-standard-library/Time_/externs_linux_and_darwin.go
@@ -1,0 +1,15 @@
+//go:build !windows
+// +build !windows
+
+package _Time
+
+import "syscall"
+
+func GetProcessCpuTimeMillis() int64 {
+	var usage syscall.Rusage
+	err := syscall.Getrusage(syscall.RUSAGE_SELF, &usage)
+	if err != nil {
+		return 0
+	}
+	return (usage.Utime.Nano() + usage.Stime.Nano()) / 1000000
+}
