+++ releases/go/smithy-dafny-standard-library/go.mod
@@ -5,3 +5,5 @@ go 1.23.0
 require github.com/dafny-lang/DafnyRuntimeGo/v4 v4.9.2
 
 require github.com/google/uuid v1.6.0
+
+require golang.org/x/sys v0.32.0 // indirect
